move users crud and fix imports
