+++ frontend/src/pages/SignUp.js
@@ -1,6 +1,6 @@
-import CreateUserForm from "#components/UsersCrud/CreateUserForm"
 import LanguageDropdown from "#components/ui/LanguageDropdown"
 import SubNavbar from "#components/ui/Navbar/SubNavbar"
+import CreateUserForm from "#components/users/UsersCrud/CreateUserForm"
 import useLocaleContext from "#context/localeContext"
 import {
   LOGO_IMAGE,
